FIX: audit 10b08 extended, AuXWitness consolidated into single shared implementation (Plonk superset moved to src/aux_witness.ts, Groth call sites renamed to loadFromPath), LineParser consolidated (Plonk stateful version removed, Plonk zkp13-16 and witness_tracker switched to shared stateless src/line_parser.ts, JSON loading centralised in src/plonk/mm_loop/load_lines.ts), changelog updated
