+++ src/aux_witness.ts
@@ -11,10 +11,17 @@ export class AuXWitness extends Struct({
   c: Fp12,
   shift_power: Field,
 }) {
-  static parse(path: string): AuXWitness {
+  static loadFromPath(path: string): AuXWitness {
     const data = fs.readFileSync(path, 'utf-8');
     const obj: AuXWitnessType = JSON.parse(data);
 
+    return new AuXWitness({
+      c: Fp12.loadFromJSON(obj.c),
+      shift_power: Field.from(obj.shift_power),
+    });
+  }
+
+  static loadFromJSON(obj: AuXWitnessType): AuXWitness {
     return new AuXWitness({
       c: Fp12.loadFromJSON(obj.c),
       shift_power: Field.from(obj.shift_power),
