+++ src/plonk/e2e_test.ts
@@ -5,7 +5,7 @@ import { VK } from './vk.js';
 import { FrC } from '../towers/fr.js';
 import { Sp1PlonkProof, deserializeProof } from './proof.js';
 import { parsePublicInputs } from './parse_pi.js';
-import { AuXWitness } from './aux_witness.js';
+import { AuXWitness } from '../aux_witness.js';
 import { createRequire } from 'module';
 const require = createRequire(import.meta.url);
 const g2_lines_required = require('./mm_loop/g2_lines.json');
