+++ src/plonk/recursion/witness_tracker.ts
@@ -26,26 +26,11 @@ import {
 import { VK } from '../vk.js';
 import { G1Affine } from '../../ec/index.js';
 import { Fp12 } from '../../towers/fp12.js';
-import { G2Line, G2LineJSON } from '../../lines/index.js';
+import { G2Line } from '../../lines/index.js';
 import { KZGLineAccumulator } from '../mm_loop/accumulate_lines.js';
 import { ATE_LOOP_COUNT } from '../../towers/consts.js';
 import { make_w27 } from '../helpers.js';
-import { createRequire } from 'module';
-const require = createRequire(import.meta.url);
-const g2_lines_required = require('../mm_loop/g2_lines.json');
-const tau_lines_required = require('../mm_loop/tau_lines.json');
-//import g2_lines_required from '../mm_loop/g2_lines.json';
-//import tau_lines_required from '../mm_loop/tau_lines.json';
-const g2_lines_path = JSON.stringify(g2_lines_required); //fs.readFileSync(`./src/plonk/mm_loop/g2_lines.json`, 'utf8');
-const tau_lines_path = JSON.stringify(tau_lines_required); //fs.readFileSync(`./src/plonk/mm_loop/tau_lines.json`, 'utf8');
-
-let parsed_g2_lines: G2LineJSON[] = JSON.parse(g2_lines_path);
-let g2_lines = parsed_g2_lines.map((g: G2LineJSON): G2Line => G2Line.fromJSON(g));
-
-let parsed_tau_lines: G2LineJSON[] = JSON.parse(tau_lines_path);
-let tau_lines = parsed_tau_lines.map(
-  (tau: G2LineJSON): G2Line => G2Line.fromJSON(tau)
-);
+import { g2_lines, tau_lines } from '../mm_loop/load_lines.js';
 
 class WitnessTracker {
   acc: Accumulator;
